backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.43
alembic==1.15.2
psycopg[binary]==3.2.13
pydantic-settings==2.8.1
python-slugify==8.0.4
pytest==8.3.5
httpx==0.28.1
